+++ woocommerce/inc/wc-loop.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * Loop
+ * WooCommerce Loop
  *
  * @package Bootscore 
  * @version 5.3.3
